Don't be overzealous with triggering full rebuilds
Because of a rule that invalidates source files if no artefact exists
yet, projects that are built for the first time would always start from
scratch, even if part of the source files had already been built
before. This is annoying when debugging, because a compiler error could
halt the build.

The reason this rule was introduced is because a change in the project
file can theoretically mean different compiler flags- and so existing
object files _should_ be invalidated. The way it was implemented
however (by comparing timestamp of artefact and project.json) isn’t the
right way.