Fix issue on MacOS where wrong libs could be loaded for environment
Bake would set an LD_LIBRARY_PATH that was not correct based on the configuration (debug, release). This was because of a complex issue.

The bake script executed export $(bake env) before actually executing the bake process. This exported the environment before running bake, which was necessary as functions like dlopen don't work if (DY)LD_LIBRARY_PATH is not set before the process is loaded.

However, this export command ignored the environment provided to the actual bake command, and could therefore end up exporting a debug environment, even though --cfg release was specified. Bake then still appended the directory to (DY)LD_LIBRARY_PATH, but because the debug directory would appear first, the debug library was loaded, and not the release library- if a debug library was found in the environment.

To overcome this issue, bake now forks itself after the configuration has been parsed and the environment has been set. This only happens on MacOS, as Linux systems pick up the LD_LIBRARY_PATH even if it has been set after the process is loaded.

As a result of this change, the bake script has been changed to no longer export the environment. This caused issues during setup, where calls like cmd("bake templates") would invoke the bake script. When executing scripts, environment variables starting with DYLD or LD can get erased (depending on security policies), and as such the setup also needed to be modified to directly use the bake executable, and not the bake script.